Add use_tty flag to password command schema
This flag will pause client input processing and UI drawing, but the client will continue processing in the background while waiting for the command to return a password to stdout. In this mode stdin is not redirected.

This enables applications to use /dev/tty to temporarily take control of the terminal to collect credentials